app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Shahar Linial <shaharlini@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.2"
pandas = "^2.2.1"
mysql-connector-python = "^8.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
